requirements.txt
absl-py==2.0.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
cftime==1.6.2
charset-normalizer==3.3.0
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==2.0.0
fastjsonschema==2.19.1
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.9.2
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.4.0
jupyter-core==5.4.0
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter-server==2.12.5
jupyter-server-terminals==0.5.2
jupyterlab==4.0.11
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.2
kiwisolver==1.4.5
llvmlite==0.41.0
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.8
netCDF4==1.6.4
networkx==3.1
<<<<<<< HEAD
=======
-e git+https://jpcurbelo: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@github.com/jpcurbelo/neuralhydrology_fork.git@3ea2639f6be641726f06a51221ca0a5505d87eb3#egg=neuralhydrology
>>>>>>> 72ba5c5a77837d309ef3436192fb6583049cbdcc
notebook-shim==0.2.3
numba==0.58.0
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.9.2.26
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
overrides==7.6.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
prometheus-client==0.19.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
terminado==0.18.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.7.0
werkzeug==3.0.0
xarray==2023.1.0
zipp==3.17.0
